refactor(deleteInvitation): change location of actions
