+++ src/pages/privileges/outlets/users/tabs/invitations/index.jsx
@@ -1,13 +1,14 @@
 import { useState } from "react";
+import { MdOutlineAssignmentTurnedIn, MdShortcut } from "react-icons/md";
 import { Table } from "../../../../../../components/data/Table";
 import { SectionMessage } from "../../../../../../components/feedback/SectionMessage";
 import { invitationEntriesDataMock } from "../../../../../../mocks/apps/privileges/invitations.mock";
 import {
   deleteInvitationUserMessageConfig,
-  invitationActionsConfig,
   invitationBreakpointsConfig,
   invitationTitlesConfig,
 } from "../../config/invitationsTable.config";
+import { DeleteInvitation } from "./DeleteInvitation";
 
 const initialMessageState = {
   show: false,
@@ -22,20 +23,39 @@ export default function InvitationsTab(props) {
   const [message, setMessage] = useState(initialMessageState);
   const [invitations, setInvitations] = useState(invitationEntriesDataMock);
 
-  const handleChangeInvitation = (invitation, action) => {
+  const invitationActionsConfig = [
+    {
+      id: 1,
+      actionName: "Complete",
+      content: <MdOutlineAssignmentTurnedIn />,
+      type: "secondary",
+    },
+    {
+      id: 2,
+      actionName: "Resend",
+      content: <MdShortcut />,
+      type: "primary",
+    },
+    {
+      id: 3,
+      actionName: "Delete",
+      content: (entry) => (
+        <DeleteInvitation invitation={entry} handleDelete={handleDelete} />
+      ),
+      type: "remove",
+    },
+  ];
+
+  const handleDelete = (invitation) => {
+    // Create fetch request here...
     let responseType = "success";
+
     try {
-      switch (action) {
-        case "removeInvitation":
-          setInvitations((prevInvitations) =>
-            prevInvitations.filter(
-              (oldInvitation) => invitation.id !== oldInvitation.id
-            )
-          );
-          break;
-        default:
-          break;
-      }
+      setInvitations((prevInvitations) =>
+        prevInvitations.filter(
+          (oldInvitation) => invitation.id !== oldInvitation.id
+        )
+      );
     } catch (error) {
       responseType = "failed";
     }
@@ -43,10 +63,19 @@ export default function InvitationsTab(props) {
     const { icon, title, description, appearance } =
       deleteInvitationUserMessageConfig[responseType];
 
+    handleShowMessage(
+      title,
+      description(invitation.username),
+      icon,
+      appearance
+    );
+  };
+
+  const handleShowMessage = (title, description, icon, appearance) => {
     setMessage({
       show: true,
       title,
-      description: description(invitation.username),
+      description,
       icon,
       appearance,
     });
@@ -64,7 +93,6 @@ export default function InvitationsTab(props) {
         actions={invitationActionsConfig}
         breakPoints={invitationBreakpointsConfig}
         filter={searchText}
-        onTriggerAction={handleChangeInvitation}
       />
       {message.show && (
         <SectionMessage
